use dashboard_json to sync dashboards

+++ datadog_sync/utils/helpers.py
@@ -89,20 +89,22 @@ def terraform_apply_resources(ctx):
 
     for resource in ctx.obj["resources"]:
         resource_dir = RESOURCE_DIR.format(resource.resource_name)
-        resource_plugin_path = resource_dir + "/.terraform/"
-        if not os.path.exists(resource_plugin_path):
-            os.mkdir(resource_plugin_path)
-        copy_tree(".terraform/", resource_plugin_path)
+        if os.path.exists(resource_dir):
+            resource_plugin_path = resource_dir + "/.terraform/"
+            if not os.path.exists(resource_plugin_path):
+                os.mkdir(resource_plugin_path)
+            copy_tree(".terraform/", resource_plugin_path)
 
-        os.chdir(resource_dir)
-        run_command(["terraform", "apply", "--auto-approve"], env)
-        # run_command(["terraform", "plan"], env)
-        copyfile(
-            "./terraform.tfstate",
-            "{}/{}/{}".format(
-                root_path,
-                DEFAULT_STATE_PATH,
-                DEFAULT_STATE_NAME.format(resource.resource_name),
-            ),
-        )
-        os.chdir(root_path)
+            os.chdir(resource_dir)
+
+            run_command(["terraform", "apply", "--auto-approve"], env)
+
+            copyfile(
+                "./terraform.tfstate",
+                "{}/{}/{}".format(
+                    root_path,
+                    DEFAULT_STATE_PATH,
+                    DEFAULT_STATE_NAME.format(resource.resource_name),
+                ),
+            )
+            os.chdir(root_path)
